Modifications related to the connection to the administration pages, checks, critical dates  & other changes
Establishment of a login form for the connection to the administration interface with a system of sessions.
Setting up of checks on id and numChapter. Modification of the links for the 1st and the last posts (homepage)
Critical dates on page Gestion des chapitres (addition of Date à déterminer for the publication date and orange text for these dates and the publication dates to come)
Other changes (number type and min=1 for the input for numChapter, redirections and minor modifications)