Changes related to new buttons Previous & Next chapter, TinyMCE, several messages & right part of the navbar on public pages  (in case of open session)
New feature of navigation with buttons Previous chapter & Next chapter. Addition of TinyMCE with some options & plugins for the chapters forms. Addition of messages : request for confirmation of the deletion of  a chapter, info for reporting a comment & info displayed on the page Visualisation de chapitre in case of a too large photo. Right part added on the public navbar in case of open session. Little simplification of code.